pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "apcore-toolkit"
version = "0.3.1"
description = "Shared scanner, schema extraction, and output toolkit for apcore framework adapters"
requires-python = ">=3.11"
readme = "README.md"
license = "Apache-2.0"
authors = [
    { name = "aipartnerup", email = "tercel.yi@gmail.com" },
]
keywords = [
    "apcore",
    "mcp",
    "scanner",
    "schema",
    "yaml",
    "pydantic",
    "openapi",
    "toolkit",
]
dependencies = [
    "apcore>=0.13.0",
    "pydantic>=2.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "apdev[dev]>=0.2.1",
]

[project.urls]
Homepage = "https://aipartnerup.com"
Repository = "https://github.com/aipartnerup/apcore-toolkit-python"
Documentation = "https://github.com/aipartnerup/apcore-toolkit-python#readme"
Issues = "https://github.com/aipartnerup/apcore-toolkit-python/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/apcore_toolkit"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
src = ["src"]
target-version = "py311"
line-length = 120

[tool.apdev]
base_package = "apcore_toolkit"
src_dir = "src"

[tool.mypy]
python_version = "3.11"
strict = true
